requirements.txt
fastapi
uvicorn
pandas
langdetect
joblib
pydantic
scikit-learn
numpy
scipy
pyjwt
cryptography
vaderSentiment
beautifulsoup4
requests
python-dotenv
pytest
jinja2
jose
nltk
python-multipart
nltk
lxml
spacy
matplotlib
seaborn
textblob
sqlalchemy
tqdm
psutil
torch
torchvision
torchaudio
transformers
sentencepiece
imblearn
emoji
xgboost
lightgbm